pyproject.toml
[tool.poetry]
name = "catswipe"
version = "0.1.0"
description = "Like tinder but for cats. But not weird. OK, not tinder... er..."
authors = ["Bradley Marques <bradleyrcmarques@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.4"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
factory-boy = "^3.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
